+++ poc/fplinterpreter.py
@@ -1,31 +1,34 @@
-import fplsourceanalyser
-import fplerror
+import poc.fplsourceanalyser
+import poc.fplsourcetransformer
+import poc.fplerror
 import tatsu
 from anytree import AnyNode, RenderTree
-from classes.AuxOutlines import AuxOutlines
-from classes.SymbolTable import SymbolTable
+from poc.classes.AuxOutlines import AuxOutlines
 
 
 class FplInterpreter(object):
-    _version = "1.1.1"
-    _errors = None
-    _source_analysis = None
-    _parser = None
-    _symbol_table_root = None
 
     def __init__(self, parser):
         self._parser = parser
         self._errors = []
         self._symbol_table_root = AnyNode(outline=AuxOutlines.root)
+        self._source_analysis = None
+        self.version = "1.1.2"
 
-    def syntax_analysis(self, theory_name, fpl_source):
+    def syntax_analysis(self, theory_name: str, fpl_source: str):
         try:
-            self._source_analysis = fplsourceanalyser.FPLSourceAnalyser(self._symbol_table_root, theory_name,
-                                                                        self._errors)
+            self._source_analysis = poc.fplsourceanalyser.FPLSourceAnalyser(self._symbol_table_root, theory_name,
+                                                                            self._errors)
             self._parser.parse(fpl_source, semantics=self._source_analysis, whitespace='')
-            # self.print_semantics()
         except tatsu.exceptions.FailedParse as ex:
-            self._source_analysis.errors.append(fplerror.FplParserError(ex, "in " + theory_name + ":" + str(ex)))
+            self._source_analysis.errors.append(poc.fplerror.FplParserError(ex, "in " + theory_name + ":" + str(ex)))
+
+    def syntax_transform(self, theory_name: str, fpl_source: str):
+        try:
+            self._syntax_transform = poc.fplsourcetransformer.FPLSourceTransformer(theory_name, self._errors)
+            self._parser.parse(fpl_source, semantics=self._syntax_transform, whitespace='')
+        except tatsu.exceptions.FailedParse as ex:
+            self._source_analysis.errors.append(poc.fplerror.FplParserError(ex, "in " + theory_name + ":" + str(ex)))
 
     def print_symbol_table(self):
         print(RenderTree(self._symbol_table_root))
@@ -33,14 +36,12 @@ class FplInterpreter(object):
     def symbol_table_to_str(self):
         return str(RenderTree(self._symbol_table_root))
 
-    def symbol_table_clear(self):
-        self._symbol_table_root.children = []
+    def clear(self):
+        self._symbol_table_root.children = tuple([])
+        self._errors.clear()
 
     def get_symbol_table_root(self):
         return self._symbol_table_root
-
-    def get_version(self):
-        return self._version
 
     def has_errors(self):
         return len(self._errors) > 0
@@ -57,10 +58,10 @@ class FplInterpreter(object):
         return self._errors
 
     def minified(self):
-        return self._source_analysis.get_minified()
+        return self._syntax_transform.get_minified()
 
     def prettyfied(self):
-        return self._source_analysis.get_prettified()
+        return self._syntax_transform.get_prettified()
 
     def print_semantics(self):
         for item in self._source_analysis.parse_list: