+++ grammar/CHANGES.md
@@ -7,3 +7,5 @@ Initial grammar
 * DerivedPredicate replaced by Predicate in PremiseOrOtherPredicate to prevent not meaningful productions like "assume qed"
 ## 1.0.3
 * SW in Alias instead of IW
+## 1.0.4
+* VariableType replaced by Type in CoordInSignature to simplify the syntax